Index generators (#17)
The index generators are used for implementing loop structures with indexing over the input and output variables in an expression. It also handles register allocation for temporary and auxiliary variables.

The IndexGenerator class is an abstract base class that the expression generators use for generating code for indexing input and output variables.

The ElementwiseIndexGenerator implements element-wise operations over input and output tensors with support for broadcasting. While it supports general broadcasting with one broadcast dimension, it has optimizations for simpler cases.