Import SyntaxNet and DRAGNN source into third-party directory (#26)
I have imported the source code for SyntaxNet and DRAGNN into //third_party/syntaxnet so we no longer have any external dependencies on this. With this change, you only need to have the Tensorflow Python library installed to build the SEMPAR trainer. All the DRAGNN ops are compiled into sempar.so, which is loaded by train.py, so you no longer need to install the DRAGNN Python package.

The C++ code is compiled against the installed Tensorflow Python library to ensure that the custom ops are compatible with the installed version of Tensorflow.

I have tested this with the newest version of Tensorflow (r1.3). You need to ensure that the Python version of protocol buffers used by Tensorflow matched the version specified in the WORKSPACE file (currently protocol buffers version 3.3).

I have fixed the initialization issues for shared libraries. The shared libraries link with //base:libinit so the module initializers are now run when the shared library is loaded. There is still an issue with using logging (i.e. using LOG(INFO) and VLOG(n)) in the module initializers. I will take a look at this later, but for now, I have just removed logging from the initialization functions.

I am using the new native Bazel rules for building protocol buffers.