Propagates Variable.ref in Flow::Eliminate()
I ran into an issue where the .ref-ness of a Variable disappeared after flow analysis.  This fixes the issue, so I guess there must have been a transformation that eliminated the output node.  I think it's reasonable because the neighboring code also propagates other Variable properties like .in and .out.