Adds Instance::SetReference() and updates DCHECKs
The SetReference() is like the Set() method but does not require a channel/connector.  As a result, the caller (not the channel) is responsible for ensuring proper alignment.  It's actually already possible to do this using "reinterpret_cast<char**>(instance.GetAddress(tensor))", so let me know if you'd rather not add a method like SetReference().

Bonus changes:
* Adds "param->name()" to DCHECK messages.  I found this was helpful for debugging.
* Adds DCHECKs on ref-ness.  I think it's true that Get() forbids ref and Set() requires ref?