Convert SEMPAR tensorflow model to Myelin flow (#33)
I have made a converter in Python for extracting a Myelin model from a SEMPAR Tensorflow model. This seems to be much simpler than the old converter written in C++.

I have also updated the parser runtime to use the new models. I now store the lexicon, commons, and action table in blobs in the flow file.