Prevent optimizing gather for ref outputs (#41)
I have added checks so we no longer optimize Gather ops when the output is a ref output. This fixes the issue in #40.

I have also added a GetData() method to Flow::Variable to make it a little easier to read constants from variables and I have added an defined() method to Shape that checks if a shape is fully defined.