Pytorch trainer for Sempar (#167)
This is a near-identical replication of the DRAGNN trainer and is written completely in python with the exception of the frame-evaluation task.

The top-level trainer is at:
sling/nlp/parser/tools/train_pytorch.py, which depends on a bunch of python files in sling/nlp/parser/trainer and sling/nlp/parser/trainer/flow. While training, it outputs a best checkpoint in the form of a Myelin flow file, which works as-is with the Myelin parser (sling/nlp/parser/tools/parse.cc).

The trainer is not optimized for speed but some optimizations can be handled in future commits.

This PR also fixes a few minor bugs:

Handling of negative indices in pyarray.cc
Handling "True" as true in flow attributes.
Adding a newline at the end of the word-vocab file, otherwise the last word is ignored.