Use LexicalEncoder in the Myelin parser (#195)
This PR substantially simplifies the Myelin parser by using the lexical encoder. Consequently, the pytorch trainer now uses a single set of embeddings for both LSTMs. This PR also contains a workaround to fix a minor discrepancy in feature extraction. This should be eventually fixed when the feature extraction is made identical across pytorch and Myelin.