Cascaded trainer architecture (#213)
A cascade is defined as a set of delegates and each final transition on a parser state is determined by a sequence of delegates. Each delegate either decides on a part (or all) of the parser action, or delegates the decision to another delegate. An example of such a cascade is the ShiftPropbankEvoke cascade, where the first delegate decides whether to SHIFT or not, a second one decides (in case of a non-SHIFT action) whether to output a non-Propbank-EVOKE action or not, and a third one decides (in case of a Propbank EVOKE) on the specific PropBank EVOKE action.

This is currently modeled as one softmax layer per delegate, and all softmax layers are connected to the feedforward unit's hidden layer. These layers are individually significantly thinner than the current monolithic softmax layer.

The cascade architecture offers the following advantages:
a) Speedup: We only invoke the softmax layers of delegates as required, and these layers are smaller. E.g. roughly half of our actions are SHIFT, which means that half the time we don't compute anything beyond the binary output ShiftOrNot delegate.

b) Expressiveness: We can add and pair custom delegates as required. Each delegate can define its own loss, e.g. ranking loss instead of softmax loss for coreference delegates. The proposed architecture goes beyond just softmax layers/losses.

c) Scalability: As we add more action types and instances, the current monolithic architecture becomes increasingly untenable. Having cascades allows us to refactor the modeling into small delegates.

This PR also makes the ShiftPropbankEvoke cascade the default for CASPAR. Without any optimizations, the forward pass with this cascade is 2.2x faster, and the backward pass is roughly 1.2x faster, leading to an overall 25% increasing in training speed for the same accuracy.