Wikicat: Enforce category_contains constraint (#351)
If a category has "category_contains" populated, then we restrict its members to belong to that class.
e.g. ("american cyclists", "category_contains", "human") and ("Lance Armstrong", "instance of", "human").

The instance_of check is done transitively using subclass-of edges, and if a category doesn't have a category_contains role, then the check is not performed.